+++ report-tests/src/test/java/eu/tsystems/mms/tic/testframework/report/testundertest/ReportTestUnderTestCorridorLow.java
@@ -25,7 +25,7 @@ import eu.tsystems.mms.tic.testframework.execution.testng.NonFunctionalAssert;
 import eu.tsystems.mms.tic.testframework.report.FailureCorridor;
 import eu.tsystems.mms.tic.testframework.report.general.TestsUnderTestGroup;
 import eu.tsystems.mms.tic.testframework.report.model.steps.TestStep;
-import pageobjects.ExitPointCreatorTestClass2;
+import io.testerra.report.test.pages.ExitPointCreatorTestClass2;
 import org.testng.Assert;
 import org.testng.annotations.Test;
